requirements.txt
# FastAPI + SQLModel Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
httpx==0.28.1
python-multipart==0.0.20
websockets==14.2
reportlab==4.2.5
aiosmtplib==3.0.2
resend==2.7.0
email-validator==2.2.0
loguru==0.7.3
supabase==2.13.0
